# public/swagger.json
{
  "openapi": "3.0.0",
  "name": "some_param",
  "info": {
    "title": "Sample API",
    "version": "1.0.0",
    "description": "This API handles multiple services like image processing, chat, file reading, and more."
  },
  "tags": [
    {
      "name": "Models",
      "description": "Operations related to test services"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "api-key"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/ImageReader": {
      "post": {
        "tags": ["Models"],
        "summary": "Read Image from URL",
        "operationId": "ApexImageReader",
        "parameters": [
          {
            "name": "image_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL of the image to be processed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Data Successful"
                    },
                    "user_id": {
                      "type": "string",
                      "description": "User ID generated from the image"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/ImageAnalyzer": {
      "post": {
        "tags": ["Models"],
        "summary": "Analyze Image",
        "operationId": "ApexImageAnalyzer",
        "parameters": [
          {
            "name": "imagePath",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL of the image to analyze"
            }
          },
          {
            "name": "question",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The question to ask based on the image content"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image analysis completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Image Analyser processed successfully"
                    },
                    "response": {
                      "type": "object",
                      "properties": {
                        "responseTime": {
                          "type": "string",
                          "description": "Time taken for the analysis"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "executionTime": {
                      "type": "integer",
                      "description": "The execution time in milliseconds"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/FileReader": {
      "post": {
        "tags": ["Models"],
        "summary": "Read File from URL",
        "operationId": "ApexFileReader",
        "parameters": [
          {
            "name": "file_url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "URL of the file to be processed"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Data Successful"
                    },
                    "user_id": {
                      "type": "string",
                      "description": "User ID generated after processing the file"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Chat": {
      "post": {
        "tags": ["Models"],
        "summary": "Chat with GPT-3.5",
        "operationId": "ApexChat",
        "parameters": [
          {
            "name": "related_content",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Content or documentation to base the chat on"
            }
          },
          {
            "name": "question",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The question to ask GPT"
            }
          },
          {
            "name": "ApiKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "API Key for authentication"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat response processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Chat processed successfully"
                    },
                    "response": {
                      "type": "string",
                      "description": "The GPT-3.5 generated response"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}